+++ src/main/java/baseball/domain/Balls.java
@@ -1,4 +1,4 @@
-package baseball.domain.ball;
+package baseball.domain;
 
 import java.util.Collection;
 import java.util.HashSet;
