+++ src/test/java/baseball/domain/BallTest.java
@@ -1,4 +1,4 @@
-package baseball.domain.ball;
+package baseball.domain;
 
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
